Split up cellularAutomataExample
